character encoding corrected for Jenkins
Once I started the json_cve_parser.py in the Jenkins machine the python interpreter failed indicating a character encoding error. Imported the parser into Notepad, exported it out, and verified UTF-8 format and Unix line endings.
Scripts that end with _json are used to run the front end in the Jenkins environment thus searching for CVE issues in the Space, Dosetrac, and Com 3 projects.