report output issues resolved.
The output report at times reported junk that was part of explicitly ignored CVE issues. This happened because some of the issue description fields used the semi colon and the code was treating the semi colon as a field separator. Changed the field separator to a variable.
Added a feature to ignore comment lines from the input file.
Restored some debug statements and commented them out.