New new new!
A new script to properly sort the existing CVE parser output files. The files contain the identifier that is a string. The string is difficult to sort directly. This script takes a single spacecom cve parser output file as input. It converts the CVE-ABCD-NNNNNN to a single number such as 2009001234. That number is easy to sort. So the final output is senisble. More updates to come.